+++ internal/sdk/models/operations/listdatabasepostgrescidrs.go
@@ -13,9 +13,9 @@ type ListDatabasePostgresCidrsRequest struct {
 	// The name of the database
 	Database string `pathParam:"style=simple,explode=false,name=database"`
 	// If provided, specifies the page offset of returned results
-	Page *float64 `default:"1" queryParam:"style=form,explode=true,name=page"`
+	Page *int64 `default:"1" queryParam:"style=form,explode=true,name=page"`
 	// If provided, specifies the number of returned results
-	PerPage *float64 `default:"25" queryParam:"style=form,explode=true,name=per_page"`
+	PerPage *int64 `default:"25" queryParam:"style=form,explode=true,name=per_page"`
 }
 
 func (l ListDatabasePostgresCidrsRequest) MarshalJSON() ([]byte, error) {
@@ -43,14 +43,14 @@ func (l *ListDatabasePostgresCidrsRequest) GetDatabase() string {
 	return l.Database
 }
 
-func (l *ListDatabasePostgresCidrsRequest) GetPage() *float64 {
+func (l *ListDatabasePostgresCidrsRequest) GetPage() *int64 {
 	if l == nil {
 		return nil
 	}
 	return l.Page
 }
 
-func (l *ListDatabasePostgresCidrsRequest) GetPerPage() *float64 {
+func (l *ListDatabasePostgresCidrsRequest) GetPerPage() *int64 {
 	if l == nil {
 		return nil
 	}
