+++ internal/sdk/models/operations/listdatabases.go
@@ -15,9 +15,9 @@ type ListDatabasesRequest struct {
 	// Search term to filter databases by name
 	Q *string `queryParam:"style=form,explode=true,name=q"`
 	// If provided, specifies the page offset of returned results
-	Page *float64 `default:"1" queryParam:"style=form,explode=true,name=page"`
+	Page *int64 `default:"1" queryParam:"style=form,explode=true,name=page"`
 	// If provided, specifies the number of returned results
-	PerPage *float64 `default:"25" queryParam:"style=form,explode=true,name=per_page"`
+	PerPage *int64 `default:"25" queryParam:"style=form,explode=true,name=per_page"`
 }
 
 func (l ListDatabasesRequest) MarshalJSON() ([]byte, error) {
@@ -45,14 +45,14 @@ func (l *ListDatabasesRequest) GetQ() *string {
 	return l.Q
 }
 
-func (l *ListDatabasesRequest) GetPage() *float64 {
+func (l *ListDatabasesRequest) GetPage() *int64 {
 	if l == nil {
 		return nil
 	}
 	return l.Page
 }
 
-func (l *ListDatabasesRequest) GetPerPage() *float64 {
+func (l *ListDatabasesRequest) GetPerPage() *int64 {
 	if l == nil {
 		return nil
 	}
